==> Method - 1/1. Data Scrapping.ipynb <==
import praw
import pandas as pd

client_id = ""
client_secret = ""
user_agent = ""
username = ""
password = ""

reddit = praw.Reddit(client_id = client_id, client_secret = client_secret, user_agent = user_agent, username = username, password = password)
subreddit = reddit.subreddit('india')

flairs = ["Politics", "Non-Political", "AskIndia", "Policy/Economy", "Business/Finance", "Science/Technology", "Scheduled", "Sports", "Food", "Photography", "CAA-NRC-NPR", "Coronavirus"]
print("Total No. Of Flairs:- ", len(flairs))

data = pd.DataFrame(columns=['RID', 'Title', 'URL', 'Score', 'Comment_Score', 'Author', 'Content', 'Adult', 'Flair'])
dictionary = {'RID': [], 'Title': [], 'URL': [], 'Score': [], 'Comment_Score': [], 'Author': [], 'Content': [], 'Adult': [], 'Flair': []}

data

for flair in flairs:
    relevant_subreddits = subreddit.search(flair,limit=1000)
    for submission in relevant_subreddits:
        if submission.link_flair_text == flair:
            dictionary['RID'].append(submission.id)
            dictionary['Title'].append(submission.title)
            dictionary['URL'].append(submission.url)
            dictionary['Score'].append(submission.score)
            dictionary['Comment_Score'].append(submission.num_comments)
            dictionary['Author'].append(submission.author)
            dictionary['Content'].append(submission.selftext)
            dictionary['Adult'].append(submission.over_18)
            dictionary['Flair'].append(submission.link_flair_text)

data = pd.DataFrame(dictionary['RID'], columns=["RID"])
data['Title'] = dictionary['Title']
data['URL'] = dictionary['URL']
data['Score'] = dictionary['Score']
data['Comment_Score'] = dictionary['Comment_Score']
data['Author'] = dictionary['Author']
data['Content'] = dictionary['Content']
data['Adult'] = dictionary['Adult']
data['Flair'] = dictionary['Flair']

data.head()

data.to_csv("Reddit Data.csv")
