==> Method - 1/2. Data Cleaning.ipynb <==
import numpy as np
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

data = pd.read_csv("Reddit Data.csv")
data.drop('Unnamed: 0', inplace=True, axis=1)
data.head()

data.describe()

data.info()

data.fillna("",inplace = True)

print("Total No. Of Null Values in RID:- ", data['RID'].isnull().sum())
print("Total No. Of Null Values in Title:- ", data['Title'].isnull().sum())
print("Total No. Of Null Values in Score:- ", data['Score'].isnull().sum())
print("Total No. Of Null Values in Comments:- ", data['Comment_Score'].isnull().sum())
print("Total No. Of Null Values in Author:- ", data['Author'].isnull().sum())
print("Total No. Of Null Values in Adult:- ", data['Adult'].isnull().sum())
print("Total No. Of Null Values in Flair:- ", data['Flair'].isnull().sum())

count = {}
for i in data['Flair']:
    if i in count:
        count[i] += 1
    else:
        count[i] = 1

sorted_x = sorted(count.items(), key=lambda kv: kv[1])
sorted_x.reverse()
print(sorted_x)

data.shape

data['Title'][0]

stop_words = set(stopwords.words("english"))

def clean_text(text):
    text = re.sub('[/(){}\[\]#\|@,"".?'':;*!$]', '', text)
    text = re.sub('[^0-9A-Za-z #+_]', '', text)
    text = str(" ".join([w for w in word_tokenize(text) if w not in stop_words]))
    return text

data['Title'] = data.apply(lambda column: clean_text(column['Title']), axis=1)
data['Content'] = data.apply(lambda column: clean_text(column['Content']), axis=1)

data['Title'][0]

data['Length_Title'] = data['Title'].apply(len)
data['Length_Content'] = data['Content'].apply(len)
data.head()

data.to_csv("Cleaned Reddit Data.csv")
