==> Method - 1/4. Modelling.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import SGDClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.neural_network import MLPClassifier
from xgboost import XGBClassifier
from sklearn.metrics import accuracy_score, classification_report
from sklearn.externals import joblib

data = pd.read_csv("Cleaned Reddit Data.csv")
data.Title = data.Title.astype(str)
data.Content = data.Content.astype(str)
data.drop("Unnamed: 0", inplace=True, axis=1)
data.head()

# Taking TITLE as Feature

x = data["Title"]
y = data["Flair"]

xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)

### 1. Random Forest Classifier

model_rforest = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', RandomForestClassifier(n_estimators=300)),
                ])
model_rforest.fit(xtrain, ytrain)

y_pred = model_rforest.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 2. Multinomial Naive Bayes Classifier

model_nb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MultinomialNB()),
                ])
model_nb.fit(xtrain, ytrain)

y_pred = model_nb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 3. Stochastic Gradient Descent Classifier

model_sgd = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', SGDClassifier()),
                ])
model_sgd.fit(xtrain, ytrain)

y_pred = model_sgd.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 4. Multi Layer Perceptron Classifier

model_mlp = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MLPClassifier()),
                ])
model_mlp.fit(xtrain, ytrain)

y_pred = model_mlp.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 5. Logistic Regression

model_log = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LogisticRegression()),
                ])
model_log.fit(xtrain, ytrain)

y_pred = model_log.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 6. XGBoost Classifier

model_xgb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', XGBClassifier()),
                ])
model_xgb.fit(xtrain, ytrain)

y_pred = model_xgb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 7. Linear Support Vector Machine Classifier

model_linearsvc = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LinearSVC()),
                ])
model_linearsvc.fit(xtrain, ytrain)

y_pred = model_linearsvc.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

# Taking Content as Feature

x = data["Content"]
y = data["Flair"]

xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)

### 1. Random Forest Classifier

model_rforest = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', RandomForestClassifier(n_estimators=300)),
                ])
model_rforest.fit(xtrain, ytrain)

y_pred = model_rforest.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 2. Multinomial Naive Bayes Classifier

model_nb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MultinomialNB()),
                ])
model_nb.fit(xtrain, ytrain)

y_pred = model_nb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 3. Stochastic Gradient Descent Classifier

model_sgd = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', SGDClassifier()),
                ])
model_sgd.fit(xtrain, ytrain)

y_pred = model_sgd.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 4. Multi Layer Perceptron Classifier

model_mlp = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MLPClassifier()),
                ])
model_mlp.fit(xtrain, ytrain)

y_pred = model_mlp.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 5. Logistic Regression

model_log = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LogisticRegression()),
                ])
model_log.fit(xtrain, ytrain)

y_pred = model_log.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 6. XGBoost Classifier

model_xgb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', XGBClassifier()),
                ])
model_xgb.fit(xtrain, ytrain)

y_pred = model_xgb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 7. Linear Support Vector Machine Classifier

model_linearsvc = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LinearSVC()),
                ])
model_linearsvc.fit(xtrain, ytrain)

y_pred = model_linearsvc.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

# Taking TITLE + CONTENT + URL as Feature

x = data["Title"] + data["Content"] + data["URL"]
y = data["Flair"]

xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)

### 1. Random Forest Classifier

model_rforest = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', RandomForestClassifier(n_estimators=300)),
                ])
model_rforest.fit(xtrain, ytrain)

y_pred = model_rforest.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 2. Multinomial Naive Bayes Classifier

model_nb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MultinomialNB()),
                ])
model_nb.fit(xtrain, ytrain)

y_pred = model_nb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 3. Stochastic Gradient Descent Classifier

model_sgd = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', SGDClassifier()),
                ])
model_sgd.fit(xtrain, ytrain)

y_pred = model_sgd.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 4. Multi Layer Perceptron Classifier

model_mlp = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', MLPClassifier()),
                ])
model_mlp.fit(xtrain, ytrain)

y_pred = model_mlp.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 5. Logistic Regression

model_log = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LogisticRegression()),
                ])
model_log.fit(xtrain, ytrain)

y_pred = model_log.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 6. XGBoost Classifier

model_xgb = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', XGBClassifier()),
                ])
model_xgb.fit(xtrain, ytrain)

y_pred = model_xgb.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

### 7. Linear Support Vector Machine Classifier

model_linearsvc = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LinearSVC()),
                ])
model_linearsvc.fit(xtrain, ytrain)

y_pred = model_linearsvc.predict(xtest)

print(classification_report(ytest, y_pred))
print(accuracy_score(ytest, y_pred))

Looking at the Accuracy Score and Classification Report of all the Models Trained above including three different types of features which are TITLE, CONTENT and TITLE + CONTENT + URL I found out that the model which has performed best based on the criteria described above is the Linear Support Vector Machine with Using TITLE + CONTENT + URL as Feature. So I am dumping that model for deployment purpose.

# Dumping The Model

x = data['Title'] + data['Content'] + data['URL']
y = data['Flair']

xtrain, xtest, ytrain, ytest = train_test_split(x, y, test_size=0.2)

model_linearsvc = Pipeline([('vect', CountVectorizer()),
                ('tfidf', TfidfTransformer()),
                ('clf', LinearSVC()),
                ])

final_model = model_linearsvc.fit(xtrain, ytrain)

joblib.dump(final_model, open('production_model.pkl', 'wb'))

# Loading and Checking The Model Performance On Random Data

model = joblib.load('production_model.pkl')

import nltk
import re
import praw
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
stop_words = set(stopwords.words("english"))

client_id = "zMvnVRM8MN_MRw"
client_secret = "-JIqik-pmfgku1BVB_2gcoxnMwM"
user_agent = "FlareReddit"
username = "psatreddit"
password = "pankaj123"

reddit = praw.Reddit(client_id = client_id, client_secret = client_secret, user_agent = user_agent, username = username, password = password)

def clean_text(text):
    text = re.sub('[/(){}\[\]#\|@,"".?'':;*!$]', '', text)
    text = re.sub('[^0-9A-Za-z #+_]', '', text)
    text = str(" ".join([w for w in word_tokenize(text) if w not in stop_words]))
    return text

def prediction(url):
    submission = reddit.submission(url = url)
    data['Title'] = str(submission.title)
    data['Content'] = str(submission.selftext)
    data['URL'] = str(submission.url)
    
    data['Title'] = clean_text(str(data['Title']))
    data['Content'] = clean_text(str(data['Content']))
    
    combined = data['Title'] + data['Content'] + data['URL']
    return model.predict(combined)[0]

prediction("https://www.reddit.com/r/india/comments/d1m9ld/iran_removes_antiindia_banners_from_pak_consulate/")
